Move helper functions to utils directory.
