+++ src/content-phishing-analysis.ts
@@ -1,10 +1,11 @@
-import { distance as levenshtein } from 'fastest-levenshtein'
 import finance from './static/finance.ts'
 import ecommerce from './static/ecommerce.ts'
 import { createApp } from 'vue'
 // @ts-ignore
 import PhishingAlert from './components/PhishingAlert.vue'
 import styles from './content-phishing-analysis.css?inline'
+import { analyzeDomain } from './utils/domainUtils.ts'
+import { estimateTokens, extractWebsiteInfo } from './utils/extractWebsiteInfo.ts'
 
 const initialPrompt = `
   You are a security expert analyzing websites for potential phishing indicators. Your task is to analyze provided website content and combine it with domain analysis results to detect phishing attempts.
@@ -73,92 +74,6 @@ type PhishingAnalysis = {
   recommendation: string
 }
 
-/**
- * Extracts the main domain from a full domain string.
- * Example: "bank.barclays.co.uk" -> "barclays.co.uk"
- *
- * @param {string} domain - Full domain including potential subdomains
- * @returns {string} Main domain without subdomains
- */
-function extractMainDomain(domain: string): string {
-  // Handle special cases for country TLDs (like .co.uk, .com.br)
-  const parts = domain.split('.')
-  if (parts.length >= 3 && parts[parts.length - 2] === 'co') {
-    // For domains like *.co.uk, return last 3 parts
-    return parts.slice(-3).join('.')
-  }
-  // For regular domains, return last 2 parts
-  return parts.slice(-2).join('.')
-}
-
-/**
- * Analyzes a domain for potential phishing attempts by comparing it against a list of legitimate domains.
- * The analysis includes Levenshtein distance comparison and suspicious variation detection.
- *
- * @param {string} currentDomain - The domain to analyze (can include 'www.' prefix)
- * @param {string[]} legitimateDomains - Array of legitimate domains to check against
- * @returns {Object} Analysis result containing:
- *   - isSuspicious {boolean} - Whether the domain is suspicious
- *   - legitimateDomain {string} - The matching legitimate domain (if found)
- *   - distance {number} - Levenshtein distance (if found)
- *   - similarity {number} - Similarity score between 0 and 1 (if found)
- */
-function analyzeDomain(
-  currentDomain: string,
-  legitimateDomains: string[]
-): {
-  isSuspicious: boolean
-  legitimateDomain?: string
-  distance?: number
-  similarity?: number
-} {
-  // Remove 'www.' prefix and normalize domain
-  const normalizedCurrentDomain = currentDomain.toLowerCase().replace(/^www\./, '')
-  const mainCurrentDomain = extractMainDomain(normalizedCurrentDomain)
-
-  // Check each legitimate domain
-  for (const originalLegitimateDomain of legitimateDomains) {
-    // Normalize legitimate domain as well
-    const legitimateDomain = originalLegitimateDomain.toLowerCase().replace(/^www\./, '')
-    const mainLegitimateDomain = extractMainDomain(legitimateDomain)
-
-    // Skip if current domain is a subdomain of legitimate domain
-    if (normalizedCurrentDomain.endsWith(`.${mainLegitimateDomain}`)) {
-      return { isSuspicious: false }
-    }
-
-    const distance = levenshtein(mainCurrentDomain, mainLegitimateDomain)
-    const similarity = 1 - distance / mainLegitimateDomain.length
-
-    // If very similar but not exact match
-    if (distance <= 2 && similarity >= 0.85 && mainCurrentDomain !== mainLegitimateDomain) {
-      return {
-        isSuspicious: true,
-        legitimateDomain: originalLegitimateDomain,
-        distance,
-        similarity,
-      }
-    }
-
-    // Check for suspicious variations (domain.contains but not exact match)
-    const legitWithoutTLD = mainLegitimateDomain.split('.')[0]
-    if (
-      mainCurrentDomain.includes(legitWithoutTLD) &&
-      mainCurrentDomain !== mainLegitimateDomain &&
-      similarity >= 0.5
-    ) {
-      return {
-        isSuspicious: true,
-        legitimateDomain: originalLegitimateDomain,
-        distance,
-        similarity,
-      }
-    }
-  }
-
-  return { isSuspicious: false }
-}
-
 function showPhishingAlert(data: any) {
   const shadowHost = document.createElement('div')
   shadowHost.setAttribute('id', 'phishing-alert-app-shadow-host')
@@ -175,193 +90,6 @@ function showPhishingAlert(data: any) {
   vueApp = createApp(PhishingAlert, { ...data })
   vueApp.mount(container)
   document.body.appendChild(shadowHost)
-}
-
-function estimateTokens(text: string) {
-  // Very rough approximation:  Assume 1 token ≈ 0.75 words (this is a highly variable estimate)
-  const words = text.split(/\s+/).filter((word: string) => word.length > 0).length // Split by whitespace, remove empty strings
-  const estimatedTokens = Math.round(words * 0.75)
-
-  // Add a small buffer for punctuation and special characters
-  const buffer = Math.round(words * 0.1) // 10% buffer - this is arbitrary
-
-  return estimatedTokens + buffer
-}
-
-function cleanText(text: string): string {
-  return text
-    .replace(/\s+/g, ' ') // Replace multiple whitespace with single space
-    .replace(/^\s+|\s+$/g, '') // Trim start and end
-    .replace(/\t/g, ' ') // Replace tabs with space
-}
-
-/**
- * Extracts website info while keeping within token limit
- * @param {Document} document - The document to analyze
- * @param {number} maxTokens - Maximum allowed tokens
- * @returns {string} Compressed website information
- */
-function extractWebsiteInfo(document: Document, maxTokens: number = 200): string {
-  const info: string[] = []
-  let currentTokens = 0
-
-  const addContent = (content: string): boolean => {
-    const tokenCount = estimateTokens(content)
-    if (currentTokens + tokenCount <= maxTokens * 0.9) {
-      info.push(content)
-      currentTokens += tokenCount
-      return true
-    }
-    return false
-  }
-
-  // Title
-  const title = cleanText(document.title)
-  if (title) {
-    if (!addContent('PAGE TITLE: ' + title)) return info.join('\n')
-  }
-
-  // Important meta tags
-  const metaSelectors = [
-    'meta[name="description"]',
-    'meta[name="keywords"]',
-    'meta[property="og:title"]',
-    'meta[property="og:description"]',
-    'meta[property="og:site_name"]',
-    'meta[name="application-name"]',
-    'meta[name="author"]',
-  ]
-  const metaTags = document.querySelectorAll(metaSelectors.join(','))
-  if (metaTags.length > 0) {
-    if (!addContent('\nMETA INFORMATION:')) return info.join('\n')
-    for (const tag of metaTags) {
-      const name = tag.getAttribute('name') || tag.getAttribute('property')
-      const content = tag.getAttribute('content')
-      if (name && content) {
-        if (!addContent(`META ${name}: ${cleanText(content)}`)) return info.join('\n')
-      }
-    }
-  }
-
-  // Forms
-  const forms = document.querySelectorAll('form')
-  if (forms.length > 0) {
-    if (!addContent('\nFORM DETAILS:')) return info.join('\n')
-    for (const form of forms) {
-      if (!addContent(`ACTION URL: ${form.action}`)) return info.join('\n')
-      const inputs = form.querySelectorAll('input[type], select, textarea')
-      for (const input of inputs) {
-        const type = input.getAttribute('type') || input.tagName.toLowerCase()
-        const name = input.getAttribute('name')
-        if (type && type !== 'hidden') {
-          if (!addContent(`INPUT FIELD: type=${type}${name ? `, name=${name}` : ''}`)) {
-            return info.join('\n')
-          }
-        }
-      }
-    }
-  }
-
-  // Headings
-  const headings = document.querySelectorAll('h1, h2, h3, h4, h5, h6')
-  if (headings.length > 0) {
-    if (!addContent('\nPAGE STRUCTURE:')) return info.join('\n')
-    for (const h of headings) {
-      const text = cleanText(h.textContent || '')
-      if (text.length > 0) {
-        if (!addContent(`${h.tagName}: ${text}`)) return info.join('\n')
-      }
-    }
-  }
-
-  // Buttons
-  const buttons = document.querySelectorAll('button, input[type="submit"], input[role="button"]')
-  if (buttons.length > 0) {
-    if (!addContent('\nINTERACTIVE ELEMENTS:')) return info.join('\n')
-    for (const btn of buttons) {
-      const text = cleanText(btn.textContent || btn.getAttribute('value') || '')
-      if (text.length > 0) {
-        if (!addContent(`BUTTON: ${text}`)) return info.join('\n')
-      }
-    }
-  }
-
-  // Links
-  const links = document.querySelectorAll('a[href]')
-  if (links.length > 0) {
-    if (!addContent('\nNAVIGATION:')) return info.join('\n')
-    for (const link of links) {
-      const href = link.getAttribute('href')
-      const text = cleanText(link.textContent || '')
-      if (text.length > 1 && href && !href.startsWith('#')) {
-        if (!addContent(`LINK: text="${text}", url=${href}`)) return info.join('\n')
-      }
-    }
-  }
-
-  // Style Information
-  if (!addContent('\nSTYLE INFORMATION:')) return info.join('\n')
-
-  // Logo
-  const logos = document.querySelectorAll('img[src*="logo"], img[alt*="logo"]')
-  for (const logo of logos) {
-    const src = logo.getAttribute('src')
-    const alt = logo.getAttribute('alt')
-    if (src || alt) {
-      if (!addContent(`LOGO: ${alt || ''}, source=${src || ''}`)) return info.join('\n')
-    }
-  }
-
-  // Background colors of main elements
-  const mainElement = document.querySelector('main') || document.body
-  const header = document.querySelector('header')
-  const computedStyle = window.getComputedStyle(mainElement)
-  const headerStyle = header ? window.getComputedStyle(header) : null
-
-  if (headerStyle?.backgroundColor && headerStyle.backgroundColor !== 'rgba(0, 0, 0, 0)') {
-    if (!addContent(`HEADER COLOR: ${headerStyle.backgroundColor}`)) return info.join('\n')
-  }
-  if (computedStyle.backgroundColor && computedStyle.backgroundColor !== 'rgba(0, 0, 0, 0)') {
-    if (!addContent(`MAIN BACKGROUND: ${computedStyle.backgroundColor}`)) return info.join('\n')
-  }
-
-  // Brand-colored elements (buttons, links)
-  const brandElements = document.querySelectorAll(
-    'a, button, .btn, [class*="primary"], [class*="brand"]'
-  )
-  const colorSet = new Set<string>()
-  brandElements.forEach((el) => {
-    const style = window.getComputedStyle(el)
-    const color = style.backgroundColor
-    if (color && color !== 'rgba(0, 0, 0, 0)') {
-      colorSet.add(color)
-    }
-  })
-  if (colorSet.size > 0) {
-    const colors = Array.from(colorSet).slice(0, 3) // Limit to top 3 colors
-    if (!addContent(`BRAND COLORS: ${colors.join(', ')}`)) return info.join('\n')
-  }
-
-  // Favicon
-  const favicon = document.querySelector('link[rel*="icon"]')
-  if (favicon) {
-    const href = favicon.getAttribute('href')
-    if (href && !addContent(`FAVICON: ${href}`)) return info.join('\n')
-  }
-
-  // Main text content
-  const mainText = document.body.querySelectorAll('p, li')
-  if (mainText.length > 0) {
-    if (!addContent('\nMAIN CONTENT:')) return info.join('\n')
-    for (const el of mainText) {
-      const text = cleanText(el.textContent || '')
-      if (text.length > 1) {
-        if (!addContent(`TEXT: ${text}`)) return info.join('\n')
-      }
-    }
-  }
-
-  return info.join('\n')
 }
 
 ;(async () => {
